+++ src/main/java/org/bytedeco/javacpp/tools/InfoMap.java
@@ -39,7 +39,8 @@ public class InfoMap extends HashMap<String,LinkedList<Info>> {
     public InfoMap(InfoMap parent) { this.parent = parent; }
 
     InfoMap parent = null;
-    static final String[] containers = { "std::deque", "std::list", "std::map", "std::queue", "std::set", "std::stack", "std::vector", "std::valarray", "std::pair" };
+    static final String[] containers = { "std::bitset", "std::deque", "std::list", "std::map", "std::queue", "std::set",
+                                         "std::stack", "std::vector", "std::valarray", "std::pair" };
     static final String[] simpleTypes = { "signed", "unsigned", "char", "short", "int", "long", "bool", "float", "double" };
     static { Arrays.sort(simpleTypes); }
     static final InfoMap defaults = new InfoMap(null)